🚧 Set the alerts on spanish, make just one request to get the services
